LayoutTests/imported/w3c/web-platform-tests/uievents/mouse/mouse_boundary_events_on_image_map-expected.txt

initial position


PASS mouse boundary events when simple over/out
PASS mouse boundary events when moved from an <area> to the same <area> shared by another <img>
FAIL mouse boundary events when the <area> is resized assert_equals: expected "[mouseover@area#area1-1, mouseenter@div#container, mouseenter@map#map1, mouseenter@area#area1-1, mousemove@area#area1-1, mouseout@area#area1-1, mouseleave@area#area1-1, mouseleave@map#map1, mouseover@img#img1, mouseenter@img#img1, mouseout@img#img1, mouseleave@img#img1, mouseleave@div#container]" but got "[mouseover@area#area1-1, mouseenter@div#container, mouseenter@map#map1, mouseenter@area#area1-1, mousemove@area#area1-1, mouseout@area#area1-1, mouseleave@area#area1-1, mouseleave@map#map1, mouseleave@div#container]"
FAIL mouse boundary events when both <area> and <img> are resized assert_equals: expected "[mouseover@area#area1-1, mouseenter@div#container, mouseenter@map#map1, mouseenter@area#area1-1, mousemove@area#area1-1, mouseout@area#area1-1, mouseleave@area#area1-1, mouseleave@map#map1, mouseover@img#img1, mouseenter@img#img1, mouseout@img#img1, mouseleave@img#img1, mouseleave@div#container]" but got "[mouseover@area#area1-1, mouseenter@div#container, mouseenter@map#map1, mouseenter@area#area1-1, mousemove@area#area1-1, mouseout@area#area1-1, mouseleave@area#area1-1, mouseleave@map#map1, mouseleave@div#container]"
PASS mouse boundary events when usemap is modified
FAIL mouse boundary events when usemap is modified and <img> is resized assert_equals: expected "[mouseover@area#area1-1, mouseenter@div#container, mouseenter@map#map1, mouseenter@area#area1-1, mousemove@area#area1-1, mouseout@area#area1-1, mouseleave@area#area1-1, mouseleave@map#map1, mouseover@area#area2-1, mouseenter@map#map2, mouseenter@area#area2-1, mouseout@area#area2-1, mouseleave@area#area2-1, mouseleave@map#map2, mouseleave@div#container]" but got "[mouseover@area#area1-1, mouseenter@div#container, mouseenter@map#map1, mouseenter@area#area1-1, mousemove@area#area1-1, mouseout@area#area1-1, mouseleave@area#area1-1, mouseleave@map#map1, mouseleave@div#container]"
PASS mouse boundary events when new <area> is available
FAIL mouse boundary events when new <area> is available and the <img> is resized assert_equals: expected "[mouseover@area#area1-1, mouseenter@div#container, mouseenter@map#map1, mouseenter@area#area1-1, mousemove@area#area1-1, mouseout@area#area1-1, mouseleave@area#area1-1, mouseover@area#area1-2, mouseenter@area#area1-2, mouseout@area#area1-2, mouseleave@area#area1-2, mouseleave@map#map1, mouseleave@div#container]" but got "[mouseover@area#area1-1, mouseenter@div#container, mouseenter@map#map1, mouseenter@area#area1-1, mousemove@area#area1-1, mouseout@area#area1-1, mouseleave@area#area1-1, mouseleave@map#map1, mouseleave@div#container]"
